Revert "clean-up code and function calls (#78)"
This reverts commit 9479e25f9f7804908e15c0d5eba8b5826bdc4955.
