+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,3 +1,13 @@
+### Assignee Tasks
+
 * [ ] added unit tests (or none needed)
 * [ ] written relevant docs (or none needed)
 * [ ] referenced any relevant issues (or none exist)
+
+### Guidelines
+
+Please add a description of this Pull Request's motivation, scope, outstanding issues or potential alternatives, reasoning behind the current solution, and any other relevant information for posterity.
+
+---
+
+_Your PR Notes Here_
