+++ client/music-player.js
@@ -1,8 +1,7 @@
 import {EventEmitter} from 'events'
+import axios from 'axios'
 import socket from './socket'
 import store, {playTrack, togglePause, seekTrack} from './store'
-
-const {dispatch, getState} = store
 
 const musicPlayerEvent = new EventEmitter()
 
@@ -10,7 +9,7 @@ const musicPlayerEvent = new EventEmitter()
  * handler for musicPlayerEvents when the player state changes
  * emit event to other socket when it is triggered by the channel owner
  */
-const handleStateChanged = playerState => {
+const handleStateChanged = (playerState, dispatch, getState) => {
   // get current channel, track and user from the state
   const {channel: {selectedChannel}, user} = getState()
   // id of the current channel participating
@@ -56,9 +55,9 @@ const resolveStateChange = (uri, paused, position) => ({
   shouldChangeTrack,
   shouldSeek
 }) =>
-  Promise.resolve(shouldChangeTrack && dispatch(playTrack(uri)))
-    .then(() => shouldTogglePlay && dispatch(togglePause(paused)))
-    .then(() => shouldSeek && dispatch(seekTrack(position)))
+  Promise.resolve(shouldChangeTrack && store.dispatch(playTrack(uri)))
+    .then(() => shouldTogglePlay && store.dispatch(togglePause(paused)))
+    .then(() => shouldSeek && store.dispatch(seekTrack(position)))
 
 /**
  * handler for when channel owner's player state changes
@@ -68,11 +67,12 @@ const resolveStateChange = (uri, paused, position) => ({
  */
 const handleStateReceived = receivedState => {
   // if received state is null, and our player is active, pause it just in case
-  if (!receivedState) return getState().player && dispatch(togglePause(true))
+  if (!receivedState)
+    return store.getState().player && store.dispatch(togglePause(true))
   // extract needed state from owner's player
   const {paused, track_window: {current_track: {uri}}, position} = receivedState
   // call the helper promise to determine the needed adjustment
-  return getChangedState(paused, uri, position, getState().player).then(
+  return getChangedState(paused, uri, position, store.getState().player).then(
     whatStateToChange =>
       resolveStateChange(uri, paused, position)(whatStateToChange)
   )
@@ -94,5 +94,41 @@ musicPlayerEvent.on('state-received', handleStateReceived)
 // listener for when user joins a channel
 // allow to catch-up to what's currently playing
 musicPlayerEvent.on('joined-channel', handleJoinChannel)
+
+// grants access token from user session. only handles successful request
+// - TODO: refreshing token, handling error
+export const getAccessToken = () =>
+  axios.get('/auth/spotify/token').then(res => res.data.accessToken)
+
+// creates player
+export const createPlayer = () =>
+  new Spotify.Player({
+    name: 'Listening Party Spotify Player',
+    getOAuthToken: callback => getAccessToken().then(callback)
+  })
+
+// returns a helper function for creating fetch call
+const makePlayRequest = (uri, id) => accessToken =>
+  fetch(`https://api.spotify.com/v1/me/player/play?device_id=${id}`, {
+    method: 'PUT',
+    body: JSON.stringify({uris: [uri]}),
+    headers: {
+      'Content-Type': 'application/json',
+      Authorization: `Bearer ${accessToken}`
+    }
+  })
+    .then(response => response)
+    .catch(console.log)
+
+/**
+ * function that makes an html request to web API and changes the track on the player
+ * @param {{string, object: {object: {function, string}}}} param0: nested object of uri and the player
+ * @returns {Spotify.Player}
+ * TODO:
+ * - add error-handling when playback fails
+ * - fetch new auth token when expired
+ */
+export const playNewUri = ({uri, player: {_options: {getOAuthToken, id}}}) =>
+  getOAuthToken(makePlayRequest(uri, id))
 
 export default musicPlayerEvent
