+++ client/store/player.js
@@ -1,5 +1,4 @@
-import musicPlayerEvent from '../music-player'
-import axios from 'axios'
+import musicPlayerEvent, {createPlayer, playNewUri} from '../music-player'
 
 const SET_PLAYER_INSTANCE = 'SET_PLAYER_INSTANCE'
 
@@ -11,42 +10,6 @@ export const setPlayerInstance = instance => ({
 // END - ACTION CREATOR
 
 // player listener
-
-// grants access token from user session. only handles successful request
-// - TODO: refreshing token, handling error
-export const getAccessToken = () =>
-  axios.get('/auth/spotify/token').then(res => res.data.accessToken)
-
-// creates player
-export const createPlayer = () =>
-  new Spotify.Player({
-    name: 'Listening Party Spotify Player',
-    getOAuthToken: callback => getAccessToken().then(callback)
-  })
-
-// returns a helper function for creating fetch call
-const makePlayRequest = (uri, id) => accessToken =>
-  fetch(`https://api.spotify.com/v1/me/player/play?device_id=${id}`, {
-    method: 'PUT',
-    body: JSON.stringify({uris: [uri]}),
-    headers: {
-      'Content-Type': 'application/json',
-      Authorization: `Bearer ${accessToken}`
-    }
-  })
-    .then(response => response)
-    .catch(console.log)
-
-/**
- * function that makes an html request to web API and changes the track on the player
- * @param {{string, object: {object: {function, string}}}} param0: nested object of uri and the player
- * @returns {Spotify.Player}
- * TODO:
- * - add error-handling when playback fails
- * - fetch new auth token when expired
- */
-export const playNewUri = ({uri, player: {_options: {getOAuthToken, id}}}) =>
-  getOAuthToken(makePlayRequest(uri, id))
 
 // THUNK CREATOR
 export const initializePlayerInstance = () => async (dispatch, getState) => {
@@ -60,7 +23,7 @@ export const initializePlayerInstance = () => async (dispatch, getState) => {
   // subscribe listeners  for when player state changes
   instance.addListener('player_state_changed', state => {
     // emit event to socket: check music-player.js
-    musicPlayerEvent.emit('state-changed', state)
+    musicPlayerEvent.emit('state-changed', state, dispatch, getState)
   })
   // listener for when device is ready
   instance.addListener('ready', device => {
